# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ValidationDecisionStatus = typeof ValidationDecisionStatus[keyof typeof ValidationDecisionStatus];


export const ValidationDecisionStatus = {
  confirmed: 'confirmed',
  held: 'held',
} as const;

export type ValidationSource = typeof ValidationSource[keyof typeof ValidationSource];


export const ValidationSource = {
  hris: 'hris',
  ats: 'ats',
} as const;

export type FieldResolutionSourceValues = {[key: string]: string | null};

export interface FieldResolution {
  /** @minLength 1 */
  fieldKey: string;
  selectedSource: ValidationSource;
  resolvedValue: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reason: string;
  sourceValues: FieldResolutionSourceValues;
}

export interface WorkforceValidationInput {
  status: ValidationDecisionStatus;
  /**
     * @maxLength 500
     * @nullable
     */
  decisionNote?: string | null;
  /** @maxItems 20 */
  decisions: FieldResolution[];
}

export type WorkforceValidationCaseStatus = typeof WorkforceValidationCaseStatus[keyof typeof WorkforceValidationCaseStatus];


export const WorkforceValidationCaseStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  held: 'held',
} as const;

export interface WorkforceValidationCase {
  employeeId: string;
  status: WorkforceValidationCaseStatus;
  /** @nullable */
  decisionNote?: string | null;
  /** @nullable */
  reviewedAt: string | null;
  decisions: FieldResolution[];
}

export interface ErrorResponse {
  error: string;
}