# lib/api-zod/src/generated/types/fieldResolution.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FieldResolutionSourceValues } from './fieldResolutionSourceValues';
import type { ValidationSource } from './validationSource';

export interface FieldResolution {
  /** @minLength 1 */
  fieldKey: string;
  selectedSource: ValidationSource;
  resolvedValue: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reason: string;
  sourceValues: FieldResolutionSourceValues;
}