# lib/api-zod/src/generated/types/validationDecisionStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ValidationDecisionStatus = typeof ValidationDecisionStatus[keyof typeof ValidationDecisionStatus];


export const ValidationDecisionStatus = {
  confirmed: 'confirmed',
  held: 'held',
} as const;