# lib/api-zod/src/generated/types/validationSource.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ValidationSource = typeof ValidationSource[keyof typeof ValidationSource];


export const ValidationSource = {
  hris: 'hris',
  ats: 'ats',
} as const;