# lib/api-zod/src/generated/types/workforceValidationCase.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FieldResolution } from './fieldResolution';
import type { WorkforceValidationCaseStatus } from './workforceValidationCaseStatus';

export interface WorkforceValidationCase {
  employeeId: string;
  status: WorkforceValidationCaseStatus;
  /** @nullable */
  decisionNote?: string | null;
  /** @nullable */
  reviewedAt: Date | null;
  decisions: FieldResolution[];
}