# lib/api-zod/src/generated/types/workforceValidationInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FieldResolution } from './fieldResolution';
import type { ValidationDecisionStatus } from './validationDecisionStatus';

export interface WorkforceValidationInput {
  status: ValidationDecisionStatus;
  /**
     * @maxLength 500
     * @nullable
     */
  decisionNote?: string | null;
  /** @maxItems 20 */
  decisions: FieldResolution[];
}